Update test result to abide by number match contract - exposed when trying to build with Java9+
git-svn-id: https://svn.apache.org/repos/asf/turbine/fulcrum/trunk@1863658 13f79535-47bb-0310-9956-ffa450edef68

+++ script/src/test/org/apache/fulcrum/script/RhinoTest.java
@@ -91,6 +91,6 @@ public class RhinoTest extends AbstractScriptTest
                         "fib(n);"; // this will be returned
 
         Number result = (Number) engine.eval(script);
-        assertEquals(55.0, result);
+        assertEquals(55, result.intValue());
     }
 }
